dev: refactored DappletsGroup
