+++ src/UIkit/NotAvailable/NotAvailable.module.css
@@ -1,0 +1,63 @@
+.fallback {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: calc(100vh - 285px);
+}
+
+.text {
+  z-index: 5;
+  font-size: 20px;
+  color: var(--textgray);
+}
+
+.central-body {
+  display: flex;
+  flex-direction: column;
+}
+
+.spin-earth-on-hover {
+  transition: ease 200s;
+  transform: rotate(-3600deg);
+}
+
+.objects img {
+  z-index: 4;
+  pointer-events: none;
+}
+
+.object-rocket {
+  position: absolute;
+  top: 75%;
+  z-index: 5;
+  pointer-events: none;
+  transform: translateX(-50px);
+  animation: rocket-movement 200s linear infinite both running;
+}
+
+.object-earth {
+  position: absolute;
+  top: 10%;
+  left: 15%;
+  z-index: 4;
+}
+
+.object-moon {
+  position: absolute;
+  top: 12%;
+  left: 25%;
+}
+
+@keyframes rocket-movement {
+  100% {
+    transform: translate(1200px, -600px);
+  }
+}
+
+@keyframes spin-earth {
+  100% {
+    transition: transform 20s;
+    transform: rotate(-360deg);
+  }
+}
