+++ src/UIkit/NotAvailable/index.ts
@@ -1,0 +1,3 @@
+import NotAvailable from './NotAvailable'
+
+export default NotAvailable
